# docs/swagger.yaml
basePath: /
definitions:
  admission.DoorEnvelope:
    properties:
      data:
        $ref: '#/definitions/admission.DoorResponse'
    type: object
  admission.DoorResponse:
    properties:
      admittedCount:
        example: 682
        type: integer
      remaining:
        example: 318
        type: integer
    type: object
  admission.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  admission.ScanEnvelope:
    properties:
      data:
        $ref: '#/definitions/admission.ScanResponse'
    type: object
  admission.ScanRequest:
    properties:
      code:
        description: |-
          Code is whatever the scanner or the keyboard produced. It is normalised
          and check-verified by the domain, so the grouped form, lower case and
          the usual O-for-zero substitutions are all accepted.
        example: 7QP5-2NFX-W78H
        type: string
    type: object
  admission.ScanResponse:
    properties:
      admitted:
        description: Admitted is the one boolean the screen needs to pick green or
          red.
        example: true
        type: boolean
      admittedAt:
        description: |-
          AdmittedAt is when the code was FIRST used, which is what the holder is
          told when it is refused as already used.
        example: "2026-09-16T21:14:03Z"
        type: string
      admittedCount:
        example: 682
        type: integer
      orderReference:
        example: A1B2C3D4
        type: string
      outcome:
        description: |-
          Outcome is the machine-readable verdict the client branches on:
          admitted, already_admitted, void, wrong_event, not_paid, unknown,
          malformed.
        enum:
        - admitted
        - already_admitted
        - void
        - wrong_event
        - not_paid
        - unknown
        - malformed
        example: admitted
        type: string
      remaining:
        description: |-
          Remaining and Admitted counters keep the door's screen live without a
          second request.
        example: 318
        type: integer
      seat:
        allOf:
        - $ref: '#/definitions/admission.SeatResponse'
        description: |-
          Seat is the reserved chair, absent for general admission.

          The most useful thing the door gains from reserved seating: somebody has
          just been told they may come in, and the next thing they ask is where to
          sit. Sent pre-formatted as well as in parts, so the screen and the email
          cannot disagree about how a chair is named.
      sequence:
        description: Sequence and Total read as "2 de 3" on a group's tickets.
        example: 2
        type: integer
      ticketTitle:
        example: Pista
        type: string
    type: object
  admission.SeatResponse:
    properties:
      label:
        description: |-
          Label is the whole thing, formatted once on the server.

          The door, the wallet, the receipt and the confirmation email all have to
          say the same thing about a chair somebody is standing in front of, and
          four copies of that format string is four chances to disagree.
        example: Plateia A · Fila K · Assento 12
        type: string
      row:
        example: K
        type: string
      seat:
        example: "12"
        type: string
      section:
        example: Plateia A
        type: string
    type: object
  admission.TicketListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/admission.TicketResponse'
        type: array
    type: object
  admission.TicketResponse:
    properties:
      admittedAt:
        type: string
      code:
        description: |-
          Code is the printed, grouped form, and QRPayload is what a QR must
          encode. Both are sent because the page draws both, and the client must
          not have to know that one is the other with the hyphens removed.
        example: 7QP5-2NFX-W78H
        type: string
      id:
        example: adm_9f2c1d8a
        type: string
      qrPayload:
        example: 7QP52NFXW78H
        type: string
      seat:
        $ref: '#/definitions/admission.SeatResponse'
      sequence:
        example: 2
        type: integer
      status:
        enum:
        - issued
        - admitted
        - void
        example: issued
        type: string
      ticketTitle:
        example: Pista
        type: string
    type: object
...
info:
  contact: {}
  description: API base para autenticação e gestão de tickets, organizada com Clean
    Architecture.
  title: Vozkot Tickets API
  version: "1.0"
paths:
...
  /api/v1/events/{id}/door:
    get:
      description: Quantos ingressos do evento já entraram e quantos ainda faltam.
        É o número que a tela da portaria mostra. Restrito ao organizador do evento
        e a administradores.
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admission.DoorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Contadores da entrada
      tags:
      - Entrada
...
  /api/v1/events/{id}/scan:
    post:
      consumes:
      - application/json
      description: 'Lê um código de ingresso, por QR ou digitado, e o consome se for
        válido. Cada código vale uma única entrada: a segunda leitura responde `already_admitted`
        e informa quando a primeira ocorreu. Um código de outro evento responde `wrong_event`,
        e um de pedido estornado responde `void`. Restrito ao organizador do evento
        e a administradores.'
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      - description: Código lido
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admission.ScanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admission.ScanEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Validar um ingresso na entrada
      tags:
      - Entrada
...
  /api/v1/orders/{id}/tickets:
    get:
      description: 'Os ingressos de um pedido pago, com o código de cada um para apresentar
        na entrada. Restrito ao comprador do pedido e a administradores: o organizador
        do evento NÃO vê estes códigos, porque quem os tem pode entrar.'
      parameters:
      - description: ID do pedido
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admission.TicketListEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Meus ingressos
      tags:
      - Entrada
  /api/v1/orders/{id}/tickets/{admissionId}/qr.png:
    get:
      description: 'O QR Code de um ingresso, em PNG, para ser exibido ou impresso.
        Endereçado pelo ID do ingresso e não pelo código: o endereço de uma imagem
        aparece no histórico do navegador e nos registros de qualquer proxy, e um
        ID sem sessão não serve para nada. Restrito ao comprador do pedido e a administradores.'
      parameters:
      - description: ID do pedido
        in: path
        name: id
        required: true
        type: string
      - description: ID do ingresso
        in: path
        name: admissionId
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admission.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Imagem do ingresso
      tags:
      - Entrada
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Use o formato: Bearer {token}. Navegadores também podem autenticar
      por cookies httpOnly.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"